[REM-333] refactor: clean up configuration routes
- api/__init__.py: removed unused user and superSu blueprints
- api/routes/config/__init__.py: updated to only include admin blueprint and added import_routes function
- api/routes/config/user/__init__.py: deleted user configuration routes
- api/routes/config/superSu/__init__.py: deleted superSu configuration routes and related functionality
- api/routes/config/superSu/bots.py: deleted superSu bots routes
- api/routes/config/superSu/client.py: deleted superSu client routes
- client/src/routes/admin/pages/UsersComponent.vue: adjusted button layout for better UI
