+++ api/routes/config/users.py
@@ -1,0 +1,166 @@
+"""Module for Admin configuration routes."""
+
+import traceback
+
+from flask_sqlalchemy import SQLAlchemy
+from quart import (
+    Response,
+    abort,
+    jsonify,
+    make_response,
+    request,
+)
+from quart import current_app as app
+from quart_jwt_extended import get_jwt_identity, jwt_required
+
+# from crawjud.forms import UserForm, UserFormEdit
+from api.models import LicensesUsers, Users
+from api.models import SuperUser as SuperUser
+
+from . import admin
+
+
+class DeleteError(Exception):
+    """Exception raised when trying to delete the user itself."""
+
+    def __init__(self, message: str) -> None:
+        """Initialize the exception."""
+        self.message = message
+
+
+class UpdateError(Exception):
+    """Exception raised when trying to update the user itself."""
+
+    def __init__(self, message: str) -> None:
+        """Initialize the exception."""
+        self.message = message
+
+
+class InsertError(Exception):
+    """Exception raised when trying to insert the user itself."""
+
+    def __init__(self, message: str) -> None:
+        """Initialize the exception."""
+        self.message = message
+
+
+def cadastro_user(form: dict) -> None:
+    """User registration.
+
+    Args:
+        form (dict): user info.
+
+    """
+    try:
+        db: SQLAlchemy = app.extensions["sqlalchemy"]
+
+        password = form.pop("password")
+
+        usr = Users(**form)
+        usr.senhacrip = password
+
+        db.session.add(usr)
+        db.session.commit()
+
+    except Exception as e:
+        raise InsertError(message=f"Erro ao inserir usuário: {e!s}") from e
+
+
+def update_user(form: dict) -> None:
+    """Update user.
+
+    Args:
+        form (dict): user info.
+
+    """
+    try:
+        db: SQLAlchemy = app.extensions["sqlalchemy"]
+
+        password: str = form.pop("password")
+
+        usr = db.session.query(Users).filter(Users.id == form["id"]).first()
+
+        if password:
+            usr.senhacrip = str(password)
+
+        db.session.commit()
+    except Exception as e:
+        raise UpdateError(message=f"Erro ao atualizar usuário: {e!s}") from e
+
+
+def delete_user(form: dict) -> None:
+    """Delete user.
+
+    Args:
+        form (dict): user info.
+
+    """
+    db: SQLAlchemy = app.extensions["sqlalchemy"]
+
+    usr = db.session.query(Users).filter(Users.id == form["id"]).first()
+
+    if usr.id == get_jwt_identity():
+        raise DeleteError(message="Não é possível deletar seu próprio usuário.")
+    db.session.delete(usr)
+    db.session.commit()
+
+
+action = {"INSERT": cadastro_user, "UPDATE": update_user, "DELETE": delete_user}
+
+
+@admin.route("/users", methods=["GET", "POST"])
+@jwt_required
+async def users() -> Response:
+    """Render the users list template.
+
+    Returns:
+        Response: HTTP response with rendered template.
+
+    """
+    try:
+        if request.method == "POST":
+            try:
+                form_data = await request.json or await request.data or await request.form
+
+                form = {}
+                form.update(form_data)
+
+                method_request: str = form.pop("method_request")
+                action.get(method_request)(form)
+
+                message = "Salvo com sucesso!"
+                if method_request == "DELETE":
+                    message = "Deletado com sucesso!"
+
+                return await make_response(jsonify(message=message), 200)
+            except (InsertError, UpdateError, DeleteError, Exception) as e:
+                return await make_response(jsonify({"message": str(e)}, 400))
+        db: SQLAlchemy = app.extensions["sqlalchemy"]
+
+        data = []
+
+        user_id: int = get_jwt_identity()
+
+        users = db.session.query(Users)
+        user = db.session.query(Users).filter(Users.id == user_id).first()
+
+        if not user.supersu:
+            users = users.join(LicensesUsers).filter_by(license_token=user.licenseusr.license_token)
+
+        database = users.all()
+
+        for item in database:
+            item_data = {
+                "id": item.id,
+                "login": item.login,
+                "nome_usuario": item.nome_usuario,
+                "email": item.email,
+            }
+
+            data.append(item_data)
+
+        return await make_response(jsonify(database=data))
+
+    except Exception as e:
+        app.logger.exception("\n".join(traceback.format_exception(e)))
+        abort(500, description="Erro interno do servidor")
